Move hellinger into utils file
